feat!: Add versioning and undo of offer events
Events should not overwrite/delete data, as it might have to be restored on undo.
Overwrites/deletes of Offers have been replaced by the following versioning pattern:

- `prisma.Offer` is the base entity with immutable data (id, offer list id, ...)
  - `currentVersionId` is the id of the current version of the offer
  - `deleted = true` means the offer is not currently in the offer list

- `prisma.OfferVersion` contains the offer data that can change when the offer is updated (gives, wants, ...)
  - `versionNumber` is an incremental version number, starting at zero
  - `prevVersionId` is the id of an older version that this version replaces, if any

On undo of an event that created a version, that version is deleted and the previous version set as the current. If there was no previous version, the `Offer` is deleted (not just marked as deleted, as the undo can never be undone).

This pattern is general and will be applied to other entities in following commits.